serving/fastapi/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
numpy==1.24.3
torch==2.0.0
torchvision==0.15.1
pydantic==1.10.7
pillow==9.5.0